requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
faiss-cpu
unstructured[docx]
pypdf
fastapi 
uvicorn 
python-multipart
python-dotenv